Fix central authentication and charge tracking

+++ software/src/modules/charge_manager/current_allocator.h
@@ -27,7 +27,7 @@ struct cm_state_v1;
 struct cm_state_v2;
 struct cm_state_v3;
 
-uint8_t charger_authorized(ChargerState &state, const CurrentAllocatorConfig *cfg);
+int16_t charger_authorized(ChargerState &state, const CurrentAllocatorConfig *cfg);
 
 int allocate_current(
         const CurrentAllocatorConfig *cfg,
